% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_func_calltree.r
\name{calltree_hmtl}
\alias{calltree_hmtl}
\alias{calltree_html}
\title{The calltree_hmtl function}
\usage{
calltree_html(map, dest)
}
\arguments{
\item{map}{the R directory of the package}

\item{dest}{the directory where the html files should be written}
}
\description{
This function creates a html website with all R functions of a package
}
\examples{
\dontrun{
 temp <- calltree_html("R", "c:\temp\wwwcalltree")
 print(temp)
}

}
\author{
Luc De Wilde
}
