% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_calltree.r
\name{calltree_html}
\alias{calltree_html}
\title{The calltree_html function}
\usage{
calltree_html(map, dest)
}
\arguments{
\item{map}{the directory of the package}

\item{dest}{the directory where the html files should be written}
}
\description{
This function creates a html website with all R functions of a package
}
\details{
The function determines the functions in the package by
looking at the top-level assignments in the R files. Only
 LEFT_ASSIGN \code{'<-'} symbols are looked at. Function calls
via \code{do.call} or function definitions via \code{setMethod}
are ignored!
}
\examples{
\dontrun{
 temp <- calltree_html("R", "c:/temp/wwwcalltree")
}

}
\author{
Luc De Wilde
}
