% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{compare_files}
\alias{compare_files}
\title{The compare_files function}
\usage{
compare_files(
  infile1,
  infile2,
  outfile = paste0(dirname(infile1), "/", basename(infile1), basename(infile2), ".diff"),
  remove.lines = NULL,
  equal.lines = 2L
)
}
\arguments{
\item{infile1}{a connection or a character vector of length 1
with the path to the first file}

\item{infile2}{a connection or a character vector of length 1
with the path to the second file}

\item{outfile}{a connection or a character vector of length 1
with the path to the output file}

\item{remove.lines}{NULL or a character vector with strings
identifying lines which should be excluded from comparison}

\item{equal.lines}{integer value > 1L, for number of
lines to be a match, default is 2L}
}
\value{
number of differences found
}
\description{
This function compares two text files and stores the differences in a
typical .diff file.
}
\examples{
filenames <- tempfile(fileext = c(".txt", ".txt"))
outfil <- file()
writeLines(c("a", "b", "c", "d", "e", "f", "g", "h", "i", "j"),
            filenames[1])
writeLines(c("0", "1", "a", "b", "c", "e", "f", "grrr", "h", "i", "j"),
            filenames[2])
compare_files(filenames[1], filenames[2], outfil, "rr")
cat(readLines(outfil), sep="\n")

compare_files(filenames[1], filenames[2], outfil, equal.lines = 3L)
cat(readLines(outfil), sep="\n")
close(outfil)

}
\author{
Luc De Wilde
}
