% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_compare.R
\name{compare_objects}
\alias{compare_objects}
\title{Compare two R objects}
\usage{
compare_objects(object1, object2, outfile = "", ignore = NULL, max.diff = 30L)
}
\arguments{
\item{object1}{the first R object}

\item{object2}{the second R object}

\item{outfile}{a connection or a character vector of length 1
with the path to the output file, if "" (default) write to
console, if FALSE output is returned, see Value.}

\item{ignore}{regular expression(s) for name of slots or elements (of a
list) to ignore in comparing}

\item{max.diff}{maximum number of differences to show, default 30L}
}
\value{
number of differences reported or, if outfile FALSE,
character vector with differences and attribute 'diff' the number of
differences found.
}
\description{
This function compares two R objects and writes the differences to a
connection or the console.
}
\examples{
lst1 <- list(a = 2, b = 2:4, c = c("a", "b"))
lst2 <- list(a = 2, b = 2:4, c = c(c1 = "a", c2 = "b"))
lst3 <- list(a = 7, b = 2:5, c = c(c1 = "a", c2 = NA_character_))
compare_objects(lst1, lst2)
compare_objects(lst2, lst3)
compare_objects(lst2, lst4 <- lst2)
compare_objects(letters, LETTERS, max.diff = 10L)
str(compare_objects(lst2, lst3, outfile = FALSE))
compare_objects(lst2, lst3, ignore = "c")

}
\author{
Luc De Wilde
}
