% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_calls.R
\name{get_calls}
\alias{get_calls}
\title{Gets a list of function calls with the named parameters}
\usage{
get_calls(files = "")
}
\arguments{
\item{files}{a character vector naming the files to search in}
}
\value{
named list with for each file a named list with for each called
function in the file a named list of named arguments with a tuple line,
position where this named argument is found
}
\description{
This function creates a list of the function call for
all functions called in .R-files with the named parameters specified.
}
\examples{
filename <- tempfile(fileext = ".R")
writeLines(c(
 "f1 <- function(a, s3list, s4obj, ...) {",
 " b <- paste(\"sum:\", s3list$teller + s4obj@teller)",
 " return(paste0(a, b))",
 "}"
), filename)
tmp <- get_calls(filename)
for (ifile in seq_along(tmp)) {
  cat(names(tmp)[ifile], "\n")
  thefile <- tmp[[ifile]]
  for (ifunc in seq_along(thefile)) {
    cat(" ", names(thefile)[ifunc], "\n")
    thefunc <- thefile[[ifunc]]
    for (ai in seq_along(thefunc)) {
      cat("   ", names(thefunc)[ai], " ", thefunc[[ai]], "\n")
    }
  }
}
unlink(filename)

}
\author{
Luc De Wilde
}
