% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_calltree.r
\name{get_func_calltree}
\alias{get_func_calltree}
\title{Create a call tree for a function in the r files of a package}
\usage{
get_func_calltree(map, func)
}
\arguments{
\item{map}{a character item naming the directory of the package}

\item{func}{name(s) of the function for which
a calltree is demanded}
}
\value{
recursive list of functions with called functions,
when the called function also appears higher in the
branch followed by ":RECURSIVE" and if the called function
has no further (package) functions
called followed by ":TERMINAL"
}
\description{
This function makes creates a call tree for a function
in the r files of a package.
}
\details{
The function determines the functions in the package by
looking at the top-level assignments in the R files. Only
LEFT_ASSIGN \code{'<-'} symbols are looked at. Function calls
via \code{do.call} or function definitions
via \code{setMethod} are ignored!
}
\examples{
\dontrun{
 temp <- get_func_calltree("R", "get_xref")
 str(temp)
}

}
\author{
Luc De Wilde
}
