% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xref.R
\name{get_xref}
\alias{get_xref}
\title{Gets a cross-reference of variables used in top-level defined functions}
\usage{
get_xref(file = "")
}
\arguments{
\item{file}{a character vector naming the files to search in}
}
\value{
named list with for each file a named list with for each found top-level
function in the file a named list with for each variable in the function a dev_var object
}
\description{
This function creates a cross-reference of the variables for
all functions defined on the top-level in .R-files.
}
\examples{
filename <- tempfile(fileext = ".R")
writeLines(c(
 "f1 <- function(a, s3list, s4obj, ...) {",
 " b <- paste(\"sum:\", s3list$teller + s4obj@teller)",
 " return(paste0(a, b))",
 "}"
), filename)
tmp <- get_xref(filename)
for (ifile in seq_along(tmp)) {
  cat(names(tmp)[ifile], "\n")
  thefile <- tmp[[ifile]]
  for (ifunc in seq_along(thefile)) {
    cat(" ", names(thefile)[ifunc], "\n")
    thefunc <- thefile[[ifunc]]
    for (v in thefunc) {
      cat("   ", v@name)
      if (v@argument) cat(" is an argument of the function.")
      cat("\n    References: ")
      for (ref in v@xrefs) {
        cat(ref@line, ":", ref@position, if (ref@modified) "*", " | ", sep = "")
      }
      cat("\n")
    }
  }
}
unlink(filename)

}
\author{
Luc De Wilde
}
