% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldwdev-package.R
\name{ldwdev-package}
\alias{ldwdev-package}
\title{ldw development utilities}
\description{
\tabular{ll}{
Package: \tab ldwdev\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2023-12-22\cr
License: \tab GPL-2\cr
}
}
