% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_calltree.r
\name{print.calltree}
\alias{print.calltree}
\title{Prints a calltree made with get_func_calltree()}
\usage{
\method{print}{calltree}(x, ...)
}
\arguments{
\item{x}{a calltree created by get_func_calltree}

\item{...}{other parameters, ignored here}
}
\description{
This function prints a call tree in a readable format.
}
\examples{
\dontrun{
 temp <- get_func_calltree("R", "get_xref")
 print(temp)
}

}
\author{
Luc De Wilde
}
