% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2pdf.R
\name{r2pdf}
\alias{r2pdf}
\title{Creates a pdf file of an R source file}
\usage{
r2pdf(source_file, dest_file, show_pdf = FALSE)
}
\arguments{
\item{source_file}{The path to the R file.}

\item{dest_file}{The path to the file to create.
If this file already exists it will be overwritten.}

\item{show_pdf}{Logical. If TRUE the PDF will be shown after completion.}
}
\value{
NULL (invisible).
}
\description{
Creates a pdf file of an R source file e.g. to print on paper.
}
\examples{
\dontrun{
   r2pdf("R/dev_util.R", "~/dev_util.pdf", TRUE)
}

}
\author{
Luc De Wilde
}
