% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_tostring.R
\name{tostring}
\alias{tostring}
\title{Creates a character representation of an R object}
\usage{
tostring(
  x,
  sep = c("array", "none", "and", "or"),
  quote = c("double", "single", "none")
)
}
\arguments{
\item{x}{The object to transform.}

\item{sep}{The type of display when lenght of x > 1L,
'array' displays the items seperated with comma's and in parentheses,
'none' is the same as 'array' but without parentheses, 'and' and
'or' is the same as 'none' but with the als comma replaced by this
conjunction. Values other than 'array' are applied only to the
highest level (if x is a list).}

\item{quote}{The type of quotes to use to surround character items.}
}
\value{
A character vector of length 1 is returned.
}
\description{
Transforms a value to a character representation e.g. for use in messages.
}
\examples{
opt <- paste0("option", 1:4)
cat(tostring(opt, "or", "single"), "\n")
cat(tostring(opt, "none", "none"), "\n")

a <- list(aaa = c(1, 2.4, 5.8), b = letters[1:5], cc = "C",
          d = c("KLERF", "nothing"))
cat(tostring(a), "\n")
# compare with toString function (in base package) :
cat(toString(a), "\n")

}
\author{
Luc De Wilde
}
