% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xref.R
\name{xref_html}
\alias{xref_html}
\title{Create html pages with functions and xrefs of variables in R files}
\usage{
xref_html(file, dest)
}
\arguments{
\item{file}{vector of filenames}

\item{dest}{the directory where the html files should be written}
}
\description{
This function creates a html page with all R functions in an R file
and the variables used therein. The resulting htmlfiles will be in the
dest directory with name \verb{R file name}.html.
}
\details{
The function determines the functions in the R files by
looking at the top-level assignments. Only
LEFT_ASSIGN \code{'<-'} symbols are looked at!
}
\examples{
\dontrun{
 temp <- xref_html(paste0("R/", c("dev_calltree", "dev_util"), ".R"),
                        "c:/temp/examplexref")
}

}
\author{
Luc De Wilde
}
