% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_compare.R
\name{compare_objects}
\alias{compare_objects}
\title{The compare_objects function}
\usage{
compare_objects(object1, object2, outfile = "", max.diff = 30L)
}
\arguments{
\item{object1}{a connection or a character vector of length 1
with the path to the first file}

\item{object2}{a connection or a character vector of length 1
with the path to the second file}

\item{outfile}{a connection or a character vector of length 1
with the path to the output file, if "" write to
console}

\item{max.diff}{maximum number of differences to show, default 30L}
}
\value{
number of differences reported
}
\description{
This function compares two R objects and writes the differences to a
connection or the console.
}
\examples{
lst1 <- list(a = 2, b = 2:4, c = c("a", "b"))
lst2 <- list(a = 2, b = 2:4, c = c(c1 = "a", c2 = "b"))
lst3 <- list(a = 7, b = 2:5, c = c(c1 = "a", c2 = NA_character_))
compare_objects(lst1, lst2)
compare_objects(lst2, lst3)
compare_objects(lst2, lst4 <- lst2)
compare_objects(letters, LETTERS, max.diff = 10L)

}
\author{
Luc De Wilde
}
