% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xref.R
\name{get_xref}
\alias{get_xref}
\title{Gets a cross-reference of variables used in top-level defined functions}
\usage{
get_xref(file = "")
}
\arguments{
\item{file}{a character vector naming the files to search in}
}
\value{
list with for each top.level function a named list with:
\itemize{
\item func__file_ the name of the file where function is found;
\item func__offset_ number of lines before the function definition;
\item the variables used and the relative line where they are used.
}
}
\description{
This function creates a cross-reference of the variables for
all functions defined on the top-level in a .R-file.
}
\examples{
filename <- tempfile(fileext = ".R")
writeLines(c(
 "f1 <- function(a, s3list, s4obj, ...) {",
 " b <- paste(\"sum:\", s3list$teller + s4obj@teller)",
 " return(paste0(a, b))",
 "}"
), filename)
tmp <- get_xref(filename)
for (j in seq_along(tmp)) {
  cat("function ", names(tmp)[j])
  tmp1 <- tmp[[j]]
  cat(" found in", tmp1$func__file_, "at offset", tmp1$func__offset_, ".\n")
  tmp1$func__file_ <- NULL
  tmp1$func__offset_ <- NULL
  for (k in seq_along(tmp1)) {
    cat("  ", names(tmp1)[k], ":",
    tostring(tmp1[[k]], "none"), "\n")
  }
}
unlink(filename)

}
\author{
Luc De Wilde
}
