% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xref.r
\name{get_xref}
\alias{get_xref}
\title{Gets a cross-reference of variables used in top-level defined functions}
\usage{
get_xref(file = "")
}
\arguments{
\item{file}{a character vector item naming the files to search in}
}
\value{
list with for each top.level function a named list with:
 'FUNC__FILE_' the name of the file where function is found
 'FUNC__OFFSET_' number of lines before the function definition
 the variables used and the relative line where they are used
}
\description{
This function makes creates a cross-reference of the variabels for
all functions defined on the top-level in a .R-file.
}
\examples{
filename <- tempfile(fileext = ".R")
writeLines(c(
 "f1 <- function(a, s3list, s4obj, ...) {",
 " b <- paste(\"sum:\", s3list$teller + s4obj@teller)",
 " return(paste0(a, b))",
 "}"
), filename)
tmp <- get_xref(filename)
for (j in seq_along(tmp)) {
  cat("function ", names(tmp)[j])
  tmp1 <- tmp[[j]]
  cat(" found in", tmp1$FUNC__FILE_, "at offset", tmp1$FUNC__OFFSET_, ".\n")
  tmp1$FUNC__FILE_ <- NULL
  tmp1$FUNC__OFFSET_ <- NULL
  for (k in seq_along(tmp1)) {
    cat("  ", names(tmp1)[k], ":",
    tostring(tmp1[[k]], "none"), "\n")
  }
}
unlink(filename)

}
\author{
Luc De Wilde
}
