% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xref.r
\name{get_xref}
\alias{get_xref}
\title{Gets a cross-reference of variables used in top-level defined functions}
\usage{
get_xref(file = "")
}
\arguments{
\item{file}{a character item naming the file to search in}
}
\value{
list with for each top.level function a list of variables used
 and where they are used
}
\description{
This function makes creates a cross-reference of the variabels for
all functions defined on the top-level in a .R-file.
}
\examples{
filename <- tempfile(fileext = ".R")
writeLines(c(
 "f1 <- function(a, s3list, s4obj, ...) {",
 " b <- paste(\"sum:\", s3list$teller + s4obj@teller)",
 " return(paste0(a, b))",
 "}"
), filename)
tmp <- get_xref(filename)
for (j in seq_along(tmp)) {
  cat("function ", names(tmp)[j], ":\n")
  tmp1 <- tmp[[j]]
  for (k in seq_along(tmp1)) {
    cat("  ", names(tmp1)[k], ":",
    tostring(tmp1[[k]], "none"), "\n")
  }
}

}
\author{
Luc De Wilde
}
