% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_xref.r
\name{xref_html}
\alias{xref_html}
\title{Create a html page with functions and xrefs of variables}
\usage{
xref_html(file, dest, title = paste("Xref", dirname(file[1L])))
}
\arguments{
\item{file}{vector of filenames to look in}

\item{dest}{the directory where the html files should be written}

\item{title}{the title for the html file}
}
\description{
This function creates a html page with all R functions in a list
of R-files and the variables used therein.
}
\details{
The function determines the functions in the R files by
looking at the top-level assignments. Only
LEFT_ASSIGN \code{'<-'} symbols are looked at!
}
\examples{
\dontrun{
 temp <- xref_html(paste0("R/", c("dev_calltree", "dev_util"), ".R"),
                        "c:/temp/examplexref")
}

}
\author{
Luc De Wilde
}
